allow training the Scraper class with an HtmlPage
The test for the Scraper class was fixed in the process and it now uses
saved data instead of making a request to a website, which changed and
broke the tests.

The encoding is guessed using w3lib.encoding when not set, instead of
defaulting to utf-8, which was likely to fail often.