allow no_content_validate for MdrExtractor
since sometimes the extract data is empty, this will make the validated
false. but we still want to add to extracted listing data to indicate
there are some data missing on the page.

also fix a problem when the annotation was added to other records
rather than seed record. fix it by propogating the annotations to
aligned elements.